--- libs/compiler/frontend_utils.h
@@ -1,35 +1,0 @@
-/*
- *	Copyright 2020 Andrey Terekhov, Maxim Menshikov
- *
- *	Licensed under the Apache License, Version 2.0 (the "License");
- *	you may not use this file except in compliance with the License.
- *	You may obtain a copy of the License at
- *
- *		http://www.apache.org/licenses/LICENSE-2.0
- *
- *	Unless required by applicable law or agreed to in writing, software
- *	distributed under the License is distributed on an "AS IS" BASIS,
- *	WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *	See the License for the specific language governing permissions and
- *	limitations under the License.
- */
-
-#pragma once
-
-#include "analyzer.h"
-
-
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-/**
- *	Read keywords from input context
- *
- *	@param	context	RuC context
- */
-void read_keywords(analyzer *context);
-
-#ifdef __cplusplus
-} /* extern "C" */
-#endif
